reorg contracts + change auctioneer to accountid in ts
