+++ .github/workflows/tests-rs-03.yml
@@ -10,5 +10,5 @@ jobs:
       - uses: actions/checkout@v4
       - name: Install and test modules
         run: |
-          cd ./contract-rs/03-owner-claims-winner-gets-nft
+          cd ./contract-rs/03-bid-with-fts
           cargo test
