--- .github/workflows/tests-rs-04.yml
@@ -1,14 +1,0 @@
-name: 04 - Tests Contract RS
-on: push
-jobs:
-  workflows:
-    strategy:
-      matrix:
-        platform: [ubuntu-latest, macos-latest]
-    runs-on: ${{ matrix.platform }}
-    steps:
-      - uses: actions/checkout@v4
-      - name: Install and test modules
-        run: |
-          cd ./contract-rs/04-ft-owner-claims-winner-gets-nft
-          cargo test
