--- .github/workflows/tests-ts-04.yml
@@ -1,19 +1,0 @@
-name: 04 - Tests Contract TS
-on: push
-jobs:
-  workflows:
-    strategy:
-      matrix:
-        platform: [ubuntu-latest, macos-latest]
-        node-version: [18, 20]
-    runs-on: ${{ matrix.platform }}
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version: ${{ matrix.node-version }}
-      - name: Install and test modules
-        run: |
-          cd ./contract-ts/04-ft-owner-claims-winner-gets-nft
-          yarn
-          yarn test
