+++ contract-rs/01-basic-auction/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "auction-contract"
-description = "Auction Example Part 1"
+description = "Auction Example Part 2"
 version = "0.1.0"
 edition = "2021"
