+++ contract-rs/01-basic-auction/src/lib.rs
@@ -14,19 +14,23 @@ pub struct Bid {
 pub struct Contract {
     highest_bid: Bid,
     auction_end_time: U64,
+    auctioneer: AccountId,
+    claimed: bool,
 }
 
 #[near]
 impl Contract {
     #[init]
-    #[private] // Only callable by the contract's account
-    pub fn init(end_time: U64) -> Self {
+    #[private] // only callable by the contract's account
+    pub fn init(end_time: U64, auctioneer: AccountId) -> Self {
         Self {
             highest_bid: Bid {
                 bidder: env::current_account_id(),
                 bid: NearToken::from_yoctonear(1),
             },
             auction_end_time: end_time,
+            claimed: false,
+            auctioneer,
         }
     }
 
@@ -58,6 +62,19 @@ impl Contract {
         Promise::new(last_bidder).transfer(last_bid)
     }
 
+    pub fn claim(&mut self) -> Promise {
+        require!(
+            env::block_timestamp() > self.auction_end_time.into(),
+            "Auction has not ended yet"
+        );
+
+        require!(!self.claimed, "Auction has already been claimed");
+        self.claimed = true;
+
+        // Transfer tokens to the auctioneer
+        Promise::new(self.auctioneer.clone()).transfer(self.highest_bid.bid)
+    }
+
     pub fn get_highest_bid(&self) -> Bid {
         self.highest_bid.clone()
     }
@@ -65,25 +82,37 @@ impl Contract {
     pub fn get_auction_end_time(&self) -> U64 {
         self.auction_end_time
     }
+
+    pub fn get_auctioneer(&self) -> AccountId {
+        self.auctioneer.clone()
+    }
+
+    pub fn get_claimed(&self) -> bool {
+        self.claimed
+    }
 }
 
-/*
- * The rest of this file holds the inline tests for the code above
- * Learn more about Rust tests: https://doc.rust-lang.org/book/ch11-01-writing-tests.html
- */
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
     fn init_contract() {
-        let contract = Contract::init(U64::from(1000));
+        let end_time: U64 = U64::from(1000);
+        let alice: AccountId = "alice.near".parse().unwrap();
+        let contract = Contract::init(end_time.clone(), alice.clone());
 
         let default_bid = contract.get_highest_bid();
         assert_eq!(default_bid.bidder, env::current_account_id());
         assert_eq!(default_bid.bid, NearToken::from_yoctonear(1));
 
-        let end_time = contract.get_auction_end_time();
-        assert_eq!(end_time, U64::from(1000));
+        let auction_end_time = contract.get_auction_end_time();
+        assert_eq!(auction_end_time, end_time);
+
+        let auctioneer = contract.get_auctioneer();
+        assert_eq!(auctioneer, alice);
+
+        let claimed = contract.get_claimed();
+        assert_eq!(claimed, false);
     }
 }
