--- contract-rs/02-owner-claims-money/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "auction-contract"
-description = "Auction Example Part 2"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-near-sdk = "5.5.0"
-
-[dev-dependencies]
-near-sdk = { version = "5.5.0", features = ["unit-testing"] }
-near-workspaces = { version = "0.14.0", features = ["unstable"] }
-tokio = { version = "1.12.0", features = ["full"] }
-serde_json = "1"
-chrono = "0.4.38"
-
-[profile.release]
-codegen-units = 1
-opt-level = "z"
-lto = true
-debug = false
-panic = "abort"
-overflow-checks = true
