--- contract-rs/02-owner-claims-money/rust-toolchain.toml
@@ -1,4 +1,0 @@
-[toolchain]
-channel = "stable"
-components = ["rustfmt"]
-targets = ["wasm32-unknown-unknown"]
