+++ contract-rs/02-winner-gets-nft/rust-toolchain.toml
