+++ contract-rs/02-winner-gets-nft/src/ext.rs
