+++ contract-rs/02-winner-gets-nft/tests/non_fungible_token.wasm
