+++ contract-rs/03-bid-with-fts/Cargo.toml
