+++ contract-rs/03-bid-with-fts/README.md
