+++ contract-rs/03-bid-with-fts/rust-toolchain.toml
