+++ contract-rs/03-bid-with-fts/src/ext.rs
