+++ contract-rs/03-bid-with-fts/src/lib.rs
