+++ contract-rs/03-bid-with-fts/tests/fungible_token.wasm
