+++ contract-rs/03-bid-with-fts/tests/non_fungible_token.wasm
