+++ contract-rs/03-bid-with-fts/tests/test_basics.rs
