+++ contract-ts/01-basic-auction/package.json
@@ -8,7 +8,7 @@
     "test": "$npm_execpath run build && ava -- ./build/auction.wasm"
   },
   "dependencies": {
-    "near-sdk-js": "2.0.0"
+    "near-sdk-js": "1.0.0"
   },
   "devDependencies": {
     "ava": "^6.1.3",
