+++ contract-ts/01-basic-auction/sandbox-test/main.ava.js
@@ -16,6 +16,7 @@ test.beforeEach(async (t) => {
 
   const alice = await root.createSubAccount("alice", { initialBalance: NEAR.parse("10 N").toString() });
   const bob = await root.createSubAccount("bob", { initialBalance: NEAR.parse("10 N").toString() });
+  const auctioneer = await root.createSubAccount("auctioneer", { initialBalance: NEAR.parse("10 N").toString() });
   const contract = await root.createSubAccount("contract", { initialBalance: NEAR.parse("10 N").toString() });
 
   // Deploy contract (input from package.json)
@@ -24,11 +25,12 @@ test.beforeEach(async (t) => {
   // Initialize contract, finishes in 1 minute
   await contract.call(contract, "init", {
     end_time: String((Date.now() + 60000) * 10 ** 6),
+    auctioneer: auctioneer.accountId,
   });
 
   // Save state for test runs, it is unique for each test
   t.context.worker = worker;
-  t.context.accounts = { alice, bob, contract };
+  t.context.accounts = { alice, bob, contract, auctioneer };
 });
 
 test.afterEach.always(async (t) => {
@@ -39,14 +41,13 @@ test.afterEach.always(async (t) => {
 });
 
 test("Test full contract", async (t) => {
-  const { alice, bob, contract } = t.context.accounts;
+  const { alice, bob, auctioneer, contract } = t.context.accounts;
 
   // Alice makes first bid
   await alice.call(contract, "bid", {}, { attachedDeposit: NEAR.parse("1 N").toString() });
   let highest_bid = await contract.view("get_highest_bid", {});
   t.is(highest_bid.bidder, alice.accountId);
   t.is(highest_bid.bid, NEAR.parse("1 N").toString());
-
   const aliceBalance = await alice.balance();
 
   // Bob makes a higher bid
@@ -59,12 +60,29 @@ test("Test full contract", async (t) => {
   const aliceNewBalance = await alice.balance();
   t.deepEqual(aliceNewBalance.available, aliceBalance.available.add(NEAR.parse("1 N")));
 
-  // Alice tries to make a bid with less NEAR than the previous 
-  await t.throwsAsync(alice.call(contract, "bid", {}, { attachedDeposit: NEAR.parse("1 N").toString() }))
+  // Alice tires to make a bid with less NEAR than the previous 
+  await t.throwsAsync(alice.call(contract, "bid", {}, { attachedDeposit: NEAR.parse("1 N").toString() }));
+
+  // Auctioneer claims auction but did not finish
+  await t.throwsAsync(auctioneer.call(contract, "claim", {}, { gas: "300000000000000" }));
 
   // Fast forward 200 blocks
   await t.context.worker.provider.fastForward(200)
 
+  const auctioneerBalance = await auctioneer.balance();
+  const available = parseFloat(auctioneerBalance.available.toHuman());
+
+  // Auctioneer claims the auction
+  await auctioneer.call(contract, "claim", {}, { gas: "300000000000000" });
+
+  // Checks that the auctioneer has the correct balance
+  const contractNewBalance = await auctioneer.balance();
+  const new_available = parseFloat(contractNewBalance.available.toHuman());
+  t.is(new_available.toFixed(2), (available + 2).toFixed(2));
+
+  // Auctioneer tries to claim the auction again
+  await t.throwsAsync(auctioneer.call(contract, "claim", {}, { gas: "300000000000000" }))
+
   // Alice tries to make a bid when the auction is over
-  await t.throwsAsync(alice.call(contract, "bid", {}, { attachedDeposit: NEAR.parse("3 N").toString() }))
+  await t.throwsAsync(alice.call(contract, "bid", {}, { attachedDeposit: NEAR.parse("1 N").toString() }));
 });
