+++ contract-ts/01-basic-auction/src/contract.ts
@@ -10,11 +10,14 @@ class Bid {
 class AuctionContract {
   highest_bid: Bid = { bidder: '', bid: BigInt(0) };
   auction_end_time: bigint = BigInt(0);
+  auctioneer: AccountId = "";
+  claimed: boolean = false;
 
   @initialize({ privateFunction: true })
-  init({ end_time}: { end_time: bigint}) {
+  init({ end_time, auctioneer}: { end_time: bigint, auctioneer: AccountId}) {
     this.auction_end_time = end_time;
     this.highest_bid = { bidder: near.currentAccountId(), bid: BigInt(1) };
+    this.auctioneer = auctioneer;
   }
 
   @call({ payableFunction: true })
@@ -33,7 +36,7 @@ class AuctionContract {
     assert(bid > lastBid, "You must place a higher bid");
 
     // Update the highest bid
-    this.highest_bid = { bidder, bid };
+    this.highest_bid = { bidder, bid }; // Save the new bid
 
     // Transfer tokens back to the last bidder
     return NearPromise.new(lastBidder).transfer(lastBid);
@@ -47,5 +50,23 @@ class AuctionContract {
   @view({})
   get_auction_end_time(): BigInt {
     return this.auction_end_time;
+  }
+
+  @view({})
+  get_auctioneer(): AccountId {
+    return this.auctioneer;
+  }
+
+  @view({})
+  get_claimed(): boolean {
+    return this.claimed;
+  }
+
+  @call({})
+  claim() {
+    assert(this.auction_end_time <= near.blockTimestamp(), "Auction has not ended yet");
+    assert(!this.claimed, "Auction has been claimed");
+    this.claimed = true;
+    return NearPromise.new(this.auctioneer).transfer(this.highest_bid.bid)
   }
 }
