+++ contract-ts/02-winner-gets-nft/sandbox-test/main.ava.js
