+++ contract-ts/02-winner-gets-nft/sandbox-test/non_fungible_token.wasm
