+++ contract-ts/02-winner-gets-nft/tsconfig.json
