+++ contract-ts/03-bid-with-fts/README.md
