+++ contract-ts/03-bid-with-fts/package.json
