+++ contract-ts/03-bid-with-fts/sandbox-test/fungible_token.wasm
