+++ contract-ts/03-bid-with-fts/sandbox-test/main.ava.js
