+++ contract-ts/03-bid-with-fts/sandbox-test/non_fungible_token.wasm
