+++ contract-ts/03-bid-with-fts/tsconfig.json
