--- contract-ts/04-ft-owner-claims-winner-gets-nft/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "auction",
-  "version": "1.0.0",
-  "license": "(MIT AND Apache-2.0)",
-  "type": "module",
-  "scripts": {
-    "build": "near-sdk-js build src/contract.ts build/auction.wasm",
-    "test": "$npm_execpath run build && ava -- ./build/auction.wasm"
-  },
-  "dependencies": {
-    "near-sdk-js": "1.0.0"
-  },
-  "devDependencies": {
-    "ava": "^6.1.3",
-    "near-workspaces": "^3.5.0",
-    "typescript": "^5.4.5"
-  },
-  "ava": {
-    "timeout": "50000",
-    "files": ["sandbox-test/*.ava.js"]
-  }
-}
