--- contract-ts/04-ft-owner-claims-winner-gets-nft/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": true,
-    "target": "ES5",
-    "noEmit": true,
-    "noImplicitAny": false,
-  },
-  "files": [
-    "src/contract.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ],
-}
